Refactor organization hierarchy tree related tests
Move org specific fixtures, utils and tests to separate modules.
