+++ actions/tests/fixtures.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import pytest
 
 from actions.models import AttributeType
