+++ actions/tests/test_api.py
@@ -1,53 +1,16 @@
 from __future__ import annotations
 
 from itertools import permutations
-from typing import Iterable, List
 
 from django.urls import reverse
 
 import pytest
 
-from aplans.tests.tree import Tree, parse_tree_string
-
-from actions.api import ActionSerializer, OrganizationSerializer
+from actions.api import ActionSerializer
 from actions.tests.factories import ActionFactory
-from kausal_common.logging.handler import get_rich_log_console
-from orgs.models import Organization
 from orgs.tests.factories import OrganizationFactory
 
 pytestmark = pytest.mark.django_db
-
-
-def orgs_to_trees(roots: Iterable[Organization], indent=0):
-    result: list[Tree] = []
-    for root in roots:
-        tree = Tree(root.name, indent)
-        for child in orgs_to_trees(root.get_children(), indent + 4):
-            tree.add_child(child)
-        result.append(tree)
-    return result
-
-
-def tree_to_org(tree: Tree, parent: Organization | None=None):
-    org = OrganizationFactory.create(name=tree.name, abbreviation=tree.name, parent=parent)
-    for child in tree.children:
-        tree_to_org(child, org)
-    return org
-
-
-@pytest.fixture()
-def org_hierarchy():
-    trees = parse_tree_string("""
-        1
-        2
-            2.1
-            2.2
-        3
-            3.1
-            3.2
-            3.3
-    """)
-    return [tree_to_org(tree) for tree in trees]
 
 
 def test_plan_api_get(api_client, plan_list_url, plan):
@@ -337,154 +300,6 @@ def test_action_bulk_serializer_reorder(plan, order):
     actions_after_save = list(plan.actions.all())
     assert actions_after_save == actions
     assert [a1.order == a2.order for a1, a2 in zip(actions_after_save, actions)]
-
-
-def assert_org_hierarchy(expected_hierarchy: str):
-    actual_roots = Organization.get_root_nodes()
-    actual = orgs_to_trees(actual_roots)
-    expected = parse_tree_string(expected_hierarchy)
-    # We could use this, but comparing the values as strings produces nicer error messages
-    # assert len(actual) == len(expected)
-    # assert all(a.equals(e) for (a, e) in zip(actual, expected))
-    actual_str = ''.join(str(tree) for tree in actual)
-    expected_str = ''.join(str(tree) for tree in expected)
-    assert actual_str == expected_str
-
-
-def update_org_hierarchy(goal_string: str):
-    # Only handles moving subtrees around for now
-    uuid_for_name: dict[str, str | None]  = {'<dummy>': None}
-    goal = Tree('<dummy>', 0)
-    for root in parse_tree_string(goal_string, reset_indent=False):
-        root.reset_indent(4)
-        goal.add_child(root)
-        for node in root.traverse():
-            uuid_for_name[node.name] = str(Organization.objects.get(name=node.name).uuid)
-    current = Tree('<dummy>', 0)
-    for child in orgs_to_trees(Organization.get_root_nodes(), 4):
-        current.add_child(child)
-    serializer_input = []
-    for goal_node in goal.traverse():
-        current_node = current.get_node(goal_node.name)
-        assert current_node
-        changed_fields = []
-        for field in ('parent', 'left_sibling'):
-            current_value = getattr(current_node, field)
-            current_value = current_value.name if current_value else None
-            goal_value = getattr(goal_node, field)
-            goal_value = goal_value.name if goal_value else None
-            if current_value != goal_value:
-                changed_fields.append(field)
-        if changed_fields:
-            node_data = OrganizationSerializer(Organization.objects.get(name=goal_node.name)).data
-            for field in changed_fields:
-                goal_value = getattr(goal_node, field)
-                goal_uuid_str = uuid_for_name[goal_value.name] if goal_value else None
-                assert node_data[field] != goal_uuid_str
-                node_data[field] = goal_uuid_str if goal_uuid_str else None
-            serializer_input.append(node_data)
-    serializer = OrganizationSerializer(many=True, data=serializer_input, instance=Organization.objects.all())
-    assert serializer.is_valid()
-    serializer.save()
-
-
-def test_organization_bulk_serializer_move_org1_after_org2(org_hierarchy):
-    expected = """
-        2
-            2.1
-            2.2
-        1
-        3
-            3.1
-            3.2
-            3.3
-    """
-    update_org_hierarchy(expected)
-    assert_org_hierarchy(expected)
-
-
-def test_organization_bulk_serializer_move_org1_after_org21(org_hierarchy):
-    expected = """
-        2
-            2.1
-            1
-            2.2
-        3
-            3.1
-            3.2
-            3.3
-    """
-    update_org_hierarchy(expected)
-    assert_org_hierarchy(expected)
-
-
-def test_organization_bulk_serializer_move_org22_after_org3(org_hierarchy):
-    expected = """
-        1
-        2
-            2.1
-        3
-            3.1
-            3.2
-            3.3
-        2.2
-    """
-    update_org_hierarchy(expected)
-    assert_org_hierarchy(expected)
-
-
-def test_organization_bulk_serializer_move_org3_below_org22(org_hierarchy):
-    expected = """
-        1
-        2
-            2.1
-            2.2
-                3
-                    3.1
-                    3.2
-                    3.3
-    """
-    update_org_hierarchy(expected)
-    assert_org_hierarchy(expected)
-
-
-@pytest.mark.parametrize('order', permutations(range(3)))
-def test_organization_bulk_serializer_move_children_of_org3(org_hierarchy, order):
-    expected = f"""
-        1
-        2
-            2.1
-            2.2
-        3
-            3.{order[0]+1}
-            3.{order[1]+1}
-            3.{order[2]+1}
-    """
-    update_org_hierarchy(expected)
-    assert_org_hierarchy(expected)
-
-
-@pytest.mark.parametrize('order', permutations(range(3)))
-def test_organization_bulk_serializer_move_roots(org_hierarchy, order):
-    subtrees = [
-        """
-        1
-        """,
-        """
-        2
-            2.1
-            2.2
-        """,
-        """
-        3
-            3.1
-            3.2
-            3.3
-        """,
-    ]
-    expected = subtrees[order[0]] + subtrees[order[1]] + subtrees[order[2]]
-    update_org_hierarchy(expected)
-    assert_org_hierarchy(expected)
 
 
 def test_category_api_get(api_client, category_list_url, category):
